Incremental update with code reorganization. Currently broken. Some debug code enabled
