+++ SCons/Tool/ninjaCommon/Rules.py
@@ -1,0 +1,58 @@
+from .Util import get_outputs, get_rule, get_inputs, get_dependencies
+
+
+def _install_action_function(_env, node):
+    """Install files using the install or copy commands"""
+    return {
+        "outputs": get_outputs(node),
+        "rule": get_rule(node, "INSTALL"),
+        "inputs": get_inputs(node),
+        "implicit": get_dependencies(node),
+    }
+
+
+def _mkdir_action_function(env, node):
+    return {
+        "outputs": get_outputs(node),
+        "rule": get_rule(node, "CMD"),
+        # implicit explicitly omitted, we translate these so they can be
+        # used by anything that depends on these but commonly this is
+        # hit with a node that will depend on all of the fake
+        # srcnode's that SCons will never give us a rule for leading
+        # to an invalid ninja file.
+        "variables": {
+            # On Windows mkdir "-p" is always on
+            "cmd": "{mkdir}".format(
+                mkdir="mkdir $out & exit 0" if env["PLATFORM"] == "win32" else "mkdir -p $out",
+            ),
+        },
+    }
+
+
+def _copy_action_function(env, node):
+    return {
+        "outputs": get_outputs(node),
+        "inputs": get_inputs(node),
+        "rule": get_rule(node, "CMD"),
+        "variables": {
+            "cmd": "$COPY $in $out",
+        },
+    }
+
+
+def _lib_symlink_action_function(_env, node):
+    """Create shared object symlinks if any need to be created"""
+    symlinks = node.check_attributes("shliblinks")
+
+    if not symlinks or symlinks is None:
+        return None
+
+    outputs = [link.get_dir().rel_path(linktgt) for link, linktgt in symlinks]
+    inputs = [link.get_path() for link, _ in symlinks]
+
+    return {
+        "outputs": outputs,
+        "inputs": inputs,
+        "rule": get_rule(node, "SYMLINK"),
+        "implicit": get_dependencies(node),
+    }
