+++ test/ninja/ninja_test_sconscripts/sconstruct_generate_and_build_cxx.py
@@ -1,0 +1,4 @@
+# DefaultEnvironment(tools=[])
+env = Environment()
+env.Tool('ninja')
+env.Program(target = 'test2', source = 'test2.cpp')
